Обновлён дизайн headerHome
