--- Frontend/src/components/headPack/aboutButton/AboutButton.tsx
@@ -1,9 +1,0 @@
-import styles from './aboutButton.module.css';
-
-export default function aboutScroll() {
-
-    return(
-        <button className={styles.AboutButton}>о нас</button>
-    )
-
-}
