--- Frontend/src/components/headPack/aboutButton/aboutButton.module.css
@@ -1,6 +1,0 @@
-.AboutButton {
-    color: #1E90FF;
-    background: none;
-    cursor: pointer;
-    font-size: 20px;
-}
