--- Frontend/src/components/headPack/createButton/createButton.module.css
@@ -1,6 +1,0 @@
-.CreateButton {
-    color: #1E90FF;
-    background: none;
-    cursor: pointer;
-    font-size: 20px;
-}
