--- Frontend/src/components/headPack/infoButton/InfoButton.tsx
@@ -1,9 +1,0 @@
-import styles from './infoScroll.module.css';
-
-export default function InfoScroll() {
-
-    return(
-        <button className={styles.InfoScroll}>информация</button>
-    )
-
-}
