--- Frontend/src/components/headPack/infoButton/infoScroll.module.css
@@ -1,7 +1,0 @@
-.InfoScroll {
-    color: #1E90FF;
-    background: none;
-    border: none;
-    cursor: pointer;
-    font-size: 20px;
-}
