+++ Frontend/src/components/headPackHome/aboutButton/AboutButton.tsx
@@ -1,0 +1,9 @@
+import styles from './aboutButton.module.css';
+
+export default function aboutScroll() {
+
+    return(
+        <button className={styles.AboutButton}>about Us</button>
+    )
+
+}
