+++ Frontend/src/components/headPackHome/aboutButton/aboutButton.module.css
@@ -1,0 +1,12 @@
+.AboutButton {
+    color: white;
+    background: none;
+    border: none;
+    cursor: pointer;
+    font-size: 16;
+    font-weight: 700;
+}
+
+.AboutButton:hover {
+  color: #b6b8b9;
+}
