+++ Frontend/src/components/headPackHome/homeHeader/homeheader.module.css
@@ -1,0 +1,49 @@
+.header {
+    width: 100%;
+    height: 94px;
+    background-color: #031525;
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+
+    /* ОСТАВЛЯЕМ ТВОИ ОТСТУПЫ */
+    padding: 20px 50px 20px 0;
+
+    /* ДЕЛАЕМ ХЕДЕР ФИКСИРОВАННЫМ */
+    position: fixed;
+    top: 0;
+    left: 0;
+    z-index: 1000;
+}
+
+.logo {
+    margin-left: 140px;   /* Как у тебя было */
+    padding-top: 0px;
+}
+
+.buttonGroup {
+    position: absolute;
+    left: 50%;
+    transform: translateX(-50%);
+
+    width: 397px;
+    height: 44px;
+
+    background: transparent;
+    border: 1px solid #1e90ff;
+    border-radius: 36px;
+
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    padding: 0 21px;
+    box-sizing: border-box;
+}
+
+.InfoScroll,
+.AboutButton {
+    padding: 10px 0;
+    min-width: 80px;
+    text-align: center;
+    color: white;
+}
