+++ Frontend/src/components/headPackHome/infoButton/InfoButton.tsx
@@ -1,0 +1,9 @@
+import styles from './infoScroll.module.css';
+
+export default function InfoScroll() {
+
+    return(
+        <button className={styles.InfoScroll}>information</button>
+    )
+
+}
