+++ Frontend/src/styles.css
@@ -6,6 +6,6 @@
 
 body {
   font-family: 'Unbounded', sans-serif;
-  background-color: white;
+  background-color: red;
   line-height: auto;
 }
